+++ common/src/main/java/net/satisfy/candlelight/core/block/entity/CEffectFoodBlockEntity.java
@@ -2,27 +2,49 @@ package net.satisfy.candlelight.core.block.entity;
 
 import com.mojang.datafixers.util.Pair;
 import net.minecraft.core.BlockPos;
+import net.minecraft.core.HolderLookup;
+import net.minecraft.core.registries.BuiltInRegistries;
+import net.minecraft.nbt.CompoundTag;
+import net.minecraft.nbt.ListTag;
 import net.minecraft.world.effect.MobEffectInstance;
-import net.minecraft.world.effect.MobEffects;
+import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 import net.satisfy.candlelight.core.registry.EntityTypeRegistry;
-import net.satisfy.farm_and_charm.core.block.entity.EffectFoodBlockEntity;
+import net.satisfy.farm_and_charm.core.item.food.EffectFoodHelper;
+import org.apache.commons.compress.utils.Lists;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
-public class CEffectFoodBlockEntity extends EffectFoodBlockEntity {
+public class CEffectFoodBlockEntity extends BlockEntity {
+    public static final String STORED_EFFECTS_KEY = "StoredEffects";
+    private List<Pair<MobEffectInstance, Float>> effects;
 
     public CEffectFoodBlockEntity(BlockPos pos, BlockState state) {
-        super(pos, state);
-        EntityTypeRegistry.EFFECT_FOOD_BLOCK_ENTITY.get();
+        super(EntityTypeRegistry.EFFECT_FOOD_BLOCK_ENTITY.get(), pos, state);
+    }
+
+    public void addEffects(List<Pair<MobEffectInstance, Float>> effects) {
+        this.effects = effects;
+    }
+
+    public List<Pair<MobEffectInstance, Float>> getEffects() {
+        return effects != null ? effects : Lists.newArrayList();
     }
 
     @Override
-    public void addEffects(List<Pair<MobEffectInstance, Float>> effects) {
-        List<Pair<MobEffectInstance, Float>> filteredEffects = effects.stream()
-                .filter(effectPair -> effectPair.getFirst().getEffect() != MobEffects.HUNGER)
-                .collect(Collectors.toList());
-        super.addEffects(filteredEffects);
+    protected void loadAdditional(CompoundTag tag, HolderLookup.Provider provider) {
+        super.loadAdditional(tag, provider);
+        this.effects = EffectFoodHelper.fromNbt(tag.getList(STORED_EFFECTS_KEY, 10));
+    }
+
+    @Override
+    protected void saveAdditional(CompoundTag tag, HolderLookup.Provider provider) {
+        super.saveAdditional(tag, provider);
+        if (effects == null) return;
+        ListTag list = new ListTag();
+        for (Pair<MobEffectInstance, Float> effect : effects) {
+            list.add(EffectFoodHelper.createNbt((short) BuiltInRegistries.MOB_EFFECT.asHolderIdMap().getId(effect.getFirst().getEffect()), effect));
+        }
+        tag.put(STORED_EFFECTS_KEY, list);
     }
 }